% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_geom_bbands.R
\name{geom_bbands}
\alias{geom_bbands}
\alias{geom_bbands_}
\title{Plot Bollinger Bands using Moving Averages}
\usage{
geom_bbands(mapping = NULL, data = NULL, position = "identity",
  na.rm = TRUE, show.legend = NA, inherit.aes = TRUE, ma_fun = SMA,
  n = 20, sd = 2, wilder = FALSE, ratio = NULL, v = 1, wts = 1:n,
  color_ma = "darkblue", color_bands = "red", alpha = 0.15,
  fill = "grey20", ...)

geom_bbands_(mapping = NULL, data = NULL, position = "identity",
  na.rm = TRUE, show.legend = NA, inherit.aes = TRUE, ma_fun = "SMA",
  n = 10, sd = 2, wilder = FALSE, ratio = NULL, v = 1, wts = 1:n,
  color_ma = "darkblue", color_bands = "red", alpha = 0.15,
  fill = "grey20", ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by [ggplot2::aes()] or
[ggplot2::aes_()]. If specified and `inherit.aes = TRUE` (the
default), it is combined with the default mapping at the top level of the
plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:

If `NULL`, the default, the data is inherited from the plot
data as specified in the call to [ggplot2::ggplot()].

A `data.frame`, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
[ggplot2::fortify()] for which variables will be created.

A `function` will be called with a single argument,
the plot data. The return value must be a `data.frame.`, and
will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If `TRUE`, silently removes `NA` values, which
typically desired for moving averages.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. [ggplot2::borders()].}

\item{ma_fun}{The function used to calculate the moving average. Seven options are
available including: SMA, EMA, WMA, DEMA, ZLEMA, VWMA, and EVWMA. The default is
`SMA`. See [TTR::SMA()] for underlying functions.}

\item{n}{Number of periods to average over. Must be between 1 and
\code{nrow(x)}, inclusive.}

\item{sd}{The number of standard deviations to use.}

\item{wilder}{logical; if \code{TRUE}, a Welles Wilder type EMA will be
calculated; see notes.}

\item{ratio}{A smoothing/decay ratio.  \code{ratio} overrides \code{wilder}
in \code{EMA}, and provides additional smoothing in \code{VMA}.}

\item{v}{The 'volume factor' (a number in [0,1]).  See Notes.}

\item{wts}{Vector of weights.  Length of \code{wts} vector must equal the
length of \code{x}, or \code{n} (the default).}

\item{color_ma, color_bands}{Select the line color to be applied for the moving
average line and the Bollinger band line.}

\item{alpha}{Used to adjust the alpha transparency for the BBand ribbon.}

\item{fill}{Used to adjust the fill color for the BBand ribbon.}

\item{...}{Other arguments passed on to [ggplot2::layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
From tidyquant
Bollinger Bands plot a range around a moving average typically two standard deviations up and down.
The `geom_bbands()` function enables plotting Bollinger Bands quickly using various moving average functions.
The moving average functions used are specified in [TTR::SMA()]
from the TTR package. Use [coord_x_date()] to zoom into specific plot regions.
The following moving averages are available:
\itemize{
   \item \strong{\href{http://www.investopedia.com/terms/s/sma.asp}{Simple moving averages (SMA)}}:
   Rolling mean over a period defined by `n`.
   \item \strong{\href{http://www.investopedia.com/terms/e/ema.asp}{Exponential moving averages (EMA)}}: Includes
   exponentially-weighted mean that gives more weight to recent observations.
   Uses `wilder` and `ratio` args.
   \item \strong{\href{http://www.investopedia.com/ask/answers/071414/whats-difference-between-moving-average-and-weighted-moving-average.asp}{Weighted moving averages (WMA)}}:
   Uses a set of weights, `wts`, to weight observations in the moving average.
   \item \strong{\href{http://www.investopedia.com/articles/trading/10/double-exponential-moving-average.asp}{Double exponential moving averages (DEMA)}}:
   Uses `v` volume factor, `wilder` and `ratio` args.
   \item \strong{\href{https://en.wikipedia.org/wiki/Zero_lag_exponential_moving_average}{Zero-lag exponential moving averages (ZLEMA)}}:
   Uses `wilder` and `ratio` args.
   \item \strong{\href{http://www.investopedia.com/articles/trading/11/trading-with-vwap-mvwap.asp}{Volume-weighted moving averages (VWMA)}}:
   Requires `volume` aesthetic.
   \item \strong{\href{http://www.motivewave.com/studies/elastic_volume_weighted_moving_average.htm}{Elastic, volume-weighted moving averages (EVWMA)}}:
   Requires `volume` aesthetic.
}
}
\section{Aesthetics}{

The following aesthetics are understood (required are in bold):
\itemize{
   \item \strong{`x`}, Typically a date
   \item \strong{`high`}, Required to be the high price
   \item \strong{`low`}, Required to be the low price
   \item \strong{`close`}, Required to be the close price
   \item `volume`, Required for VWMA and EVWMA
   \item `colour`, Affects line colors
   \item `fill`, Affects ribbon fill color
   \item `alpha`, Affects ribbon alpha value
   \item `group`
   \item `linetype`
   \item `size`
}
}

\examples{
# Load libraries
library(tidyquant)

AAPL <- tq_get("AAPL")

# SMA
AAPL \%>\%
    ggplot(aes(x = date, y = close)) +
    geom_line() +           # Plot stock price
    geom_bbands(aes(high = high, low = low, close = close), ma_fun = SMA, n = 50) +
    coord_x_date(xlim = c(today() - years(1), today()), ylim = c(80, 130))


# EMA
AAPL \%>\%
   ggplot(aes(x = date, y = close)) +
   geom_line() +           # Plot stock price
   geom_bbands(aes(high = high, low = low, close = close),
                  ma_fun = EMA, wilder = TRUE, ratio = NULL, n = 50) +
   coord_x_date(xlim = c(today() - years(1), today()), ylim = c(80, 130))


# VWMA
AAPL \%>\%
    ggplot(aes(x = date, y = close)) +
    geom_line() +           # Plot stock price
    geom_bbands(aes(high = high, low = low, close = close, volume = volume),
                   ma_fun = VWMA, n = 50) +
    coord_x_date(xlim = c(today() - years(1), today()), ylim = c(80, 130))
}
\seealso{
See individual modeling functions for underlying parameters:
\itemize{
   \item [TTR::SMA()] for simple moving averages
   \item [TTR::EMA()] for exponential moving averages
   \item [TTR::WMA()] for weighted moving averages
   \item [TTR::DEMA()] for double exponential moving averages
   \item [TTR::ZLEMA()] for zero-lag exponential moving averages
   \item [TTR::VWMA()] for volume-weighted moving averages
   \item [TTR::EVWMA()] for elastic, volume-weighted moving averages
   \item [coord_x_date()] for zooming into specific regions of a plot
}
}
