% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_geom_chart.R
\name{geom_chart}
\alias{geom_chart}
\alias{geom_barchart}
\alias{geom_candlestick}
\title{Plot Financial Charts in ggplot2}
\usage{
geom_barchart(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, color_up = "darkblue", color_down = "red",
  fill_up = "darkblue", fill_down = "red", ...)

geom_candlestick(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, color_up = "darkblue", color_down = "red",
  fill_up = "darkblue", fill_down = "red", ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by [ggplot2::aes()] or
[ggplot2::aes_()]. If specified and `inherit.aes = TRUE` (the
default), it is combined with the default mapping at the top level of the
plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:

If `NULL`, the default, the data is inherited from the plot
data as specified in the call to [ggplot2::ggplot()].

A `data.frame`, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
[ggplot2::fortify()] for which variables will be created.

A `function` will be called with a single argument,
the plot data. The return value must be a `data.frame.`, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If `TRUE`, silently removes `NA` values, which
typically desired for moving averages.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. [ggplot2::borders()].}

\item{color_up, color_down}{Select colors to be applied based on price movement
from open to close. If close >= open, `color_up` is used. Otherwise,
`color_down` is used. The default is "darkblue" and "red", respectively.}

\item{fill_up, fill_down}{Select fills to be applied based on price movement
from open to close. If close >= open, `fill_up` is used. Otherwise,
`fill_down` is used. The default is "darkblue" and "red", respectively.
Only affects `geom_candlestick`.}

\item{...}{Other arguments passed on to [ggplot2::layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
From tidyquant
Financial charts provide visual cues to open, high, low, and close prices.
Use [coord_x_date()] to zoom into specific plot regions.
The following financial chart geoms are available:
\itemize{
   \item \strong{\href{http://www.investopedia.com/terms/b/barchart.asp}{Bar Chart}}
   \item \strong{\href{http://www.investopedia.com/terms/c/candlestick.asp}{Candlestick Chart}}
}
}
\section{Aesthetics}{

The following aesthetics are understood (required are in bold):
\itemize{
   \item \strong{`x`}, Typically a date
   \item \strong{`open`}, Required to be the open price
   \item \strong{`high`}, Required to be the high price
   \item \strong{`low`}, Required to be the low price
   \item \strong{`close`}, Required to be the close price
   \item `alpha`
   \item `group`
   \item `linetype`
   \item `size`
}
}

\examples{
# Load libraries
library(tidyquant)

AAPL <- tq_get("AAPL")

# Bar Chart
AAPL \%>\%
    ggplot(aes(x = date, y = close)) +
    geom_barchart(aes(open = open, high = high, low = low, close = close)) +
    geom_ma(color = "darkgreen") +
    coord_x_date(xlim = c(today() - weeks(6), today()),
                 ylim = c(100, 130))

# Candlestick Chart
AAPL \%>\%
    ggplot(aes(x = date, y = close)) +
    geom_candlestick(aes(open = open, high = high, low = low, close = close)) +
    geom_ma(color = "darkgreen") +
    coord_x_date(xlim = c(today() - weeks(6), today()),
                 ylim = c(100, 130))
}
\seealso{
See individual modeling functions for underlying parameters:
\itemize{
   \item [geom_ma()] for adding moving averages to ggplots
   \item [geom_bbands()] for adding Bollinger Bands to ggplots
   \item [coord_x_date()] for zooming into specific regions of a plot
}
}
